update default registry matcher
Not all addons come from k8s.gcr.io (thanks coredns), so we have to change
the matcher to back reference our default vs blindly assuming k8s.gcr.io.

The sed -E may look funny, but it's synonymous with '-r' on everything but a
mac (apparently), and we need it to employ extended regex for backrefs.