Only set --cluster-name if value is provided
To ensure that the snap works with earlier versions of the charm code
which do not send the `cluster-tag` config, we should only include the
`--cluster-name` param if we have a value to give it. If the value is
not provided, we leave off the option and allow it to fall back to the
old behavior of using the upstream default.